refactor(lib): extract path/config helpers into paths module (#217)
Move the contiguous `resolve_*` / `expand_tilde` / `appname` block (plus the
adjacent config-flag reads and update-log recording helpers) out of the
~9.2k-line lib.rs into a focused `src/paths.rs`. All items are pub(crate);
lib.rs glob-imports them so the many bare call sites are unchanged, and
re-exports the two doctor-facing `*_public` wrappers so `crate::*_public`
keeps resolving for src/doctor.rs.

Behaviour-preserving: pure code motion, no logic changes. First increment of
the per-command/per-concern decomposition tracked in #217.

cargo fmt --check + clippy -D warnings + test (805) + --doc all green.
